+++ app/src/main/java/com/example/zibook/feature_book/domain/use_case/GetBookById.kt
@@ -8,7 +8,7 @@ class GetBookById(
     private val repository: BookRepository
 ) {
 
-    suspend operator fun invoke(id: Int): Book? {
+    suspend operator fun invoke(id: Long): Book? {
         return repository.getBookById(id)
     }
 }
